Update 02_textured_raycast.c
It was drawing the new image on top of the previous image.. more than one image was displayed each time the key was pressed. 
so I did a re-initialize to the info->buf array each time it enters the calc function so it can store only the colors of the new image. And I added mlx_clear_window so that each time key is pressed it will clear the window to draw the new image. It's now working with me, I came here to change it manually .. if it didn't work please tell me I will push the changes that I have made to the repo. I hope you accept my pull request ^-^